# BackEnd/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StoryX API",
    "version": "1.0.0",
    "description": "StoryX Backend API - AI-powered animation video generation platform",
    "contact": {
      "name": "StoryX Team"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Success message"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "AIModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "llm",
              "video",
              "image",
              "tts"
            ]
          },
          "baseUrl": {
            "type": "string"
          },
          "provider": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/ai-models/providers": {
      "get": {
        "summary": "获取所有AI提供商",
        "description": "获取所有活跃的AI服务提供商列表",
        "tags": [
          "AI Models"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "成功返回提供商列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string"
                              },
                              "displayName": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "website": {
                                "type": "string"
                              },
                              "logoUrl": {
                                "type": "string"
                              },
                              "isActive": {
                                "type": "boolean"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建新提供商",
        "description": "创建一个新的AI服务提供商",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "displayName"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "提供商名称（唯一标识）"
                  },
                  "displayName": {
                    "type": "string",
                    "description": "显示名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "提供商描述"
                  },
                  "website": {
                    "type": "string",
                    "description": "官网链接"
                  },
                  "logoUrl": {
                    "type": "string",
                    "description": "Logo链接"
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "提供商创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        }
      }
    },
    "/api/ai-models/providers/all": {
      "get": {
        "summary": "获取所有提供商（包括未启用的）",
        "description": "获取所有AI服务提供商列表，包括未启用的",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回提供商列表"
          }
        }
      }
    },
    "/api/ai-models/providers/{id}": {
      "get": {
        "summary": "获取提供商详情",
        "description": "根据提供商ID获取详细信息",
        "tags": [
          "AI Models"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回提供商详情"
          },
          "404": {
            "description": "提供商不存在"
          }
        }
      },
      "put": {
        "summary": "更新提供商",
        "description": "更新提供商信息",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "logoUrl": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "提供商更新成功"
          },
          "404": {
            "description": "提供商不存在"
          }
        }
      },
      "delete": {
        "summary": "删除提供商",
        "description": "删除提供商（如果有关联的模型，将无法删除）",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "提供商删除成功"
          },
          "400": {
            "description": "提供商下还有模型，无法删除"
          },
          "404": {
            "description": "提供商不存在"
          }
        }
      }
    },
...
    "/api/ai-models/type/{type}": {
      "get": {
        "summary": "按类型获取模型",
        "description": "根据类型获取AI模型列表（llm、video、image、tts）",
        "tags": [
          "AI Models"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "type",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "llm",
                "video",
                "image",
                "tts"
              ]
            },
            "description": "模型类型"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AIModel"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "无效的模型类型",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-models/{id}": {
      "get": {
        "summary": "获取模型详情",
        "description": "根据模型ID获取AI模型的详细信息",
        "tags": [
          "AI Models"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "模型ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回模型详情",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AIModel"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "模型不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新模型",
        "description": "更新模型信息",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "providerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "llm",
                      "video",
                      "image",
                      "tts"
                    ]
                  },
                  "category": {
                    "type": "string"
                  },
                  "baseUrl": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "requiresKey": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "模型更新成功"
          },
          "404": {
            "description": "模型不存在"
          }
        }
      },
      "delete": {
        "summary": "删除模型",
        "description": "删除AI模型",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "模型删除成功"
          },
          "404": {
            "description": "模型不存在"
          }
        }
      }
    },
    "/api/ai-models/all": {
      "get": {
        "summary": "获取所有模型（包括未启用的）",
        "description": "获取所有AI模型列表，包括未启用的",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回模型列表"
          }
        }
      }
    },
    "/api/ai-models": {
      "post": {
        "summary": "创建新模型",
        "description": "创建一个新的AI模型",
        "tags": [
          "AI Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "providerId",
                  "name",
                  "displayName",
                  "type"
                ],
                "properties": {
                  "providerId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "提供商ID"
                  },
                  "name": {
                    "type": "string",
                    "description": "模型名称（唯一标识）"
                  },
                  "displayName": {
                    "type": "string",
                    "description": "显示名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "模型描述"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "llm",
                      "video",
                      "image",
                      "tts"
                    ],
                    "description": "模型类型"
                  },
                  "category": {
                    "type": "string",
                    "description": "分类"
                  },
                  "baseUrl": {
                    "type": "string",
                    "description": "API基路径"
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  },
                  "requiresKey": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "模型创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        }
      }
    },
...
  },
...
}